Hand evaluation
Uses the TwoPlusTwoHandEvaluator: https://github.com/tangentforks/TwoPlusTwoHandEvaluator

`HandRanks.dat` is a large file - ~120mb. Instead of distributing it, we should generate it from source.

For binary builds, we generate the file and copy it alongside the binary.
For tests, we look for the file in the `two-plus-two-hand-evaluator` directory. The file can be generated by calling `make` in this directory.